chatbot/requirements.txt
langchain
langchain-core
langchain-community
langchain-ollama
langchain-google-genai
streamlit
python-dotenv